+++ .env
@@ -1,0 +1,1 @@
+REACT_APP_GA_ID="G-HGYTWCP3J6"
