+++ package.json
@@ -22,6 +22,7 @@
     "clsx": "^1.2.1",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
+    "react-ga4": "^1.4.1",
     "react-router-dom": "^6.3.0",
     "scrollreveal": "^4.0.9"
   },
@@ -39,8 +40,8 @@
     "eslint-plugin-react-hooks": "^4.5.0",
     "prettier": "^2.6.2",
     "prettier-plugin-tailwindcss": "^0.1.11",
-    "tailwindcss": "^3.1.5",
     "react-scripts": "5.0.1",
+    "tailwindcss": "^3.1.5",
     "web-vitals": "^2.1.4"
   },
   "eslintConfig": {
