+++ src/routes/how-it-works.jsx
@@ -1,8 +1,10 @@
-import React from "react";
-
+import React, { useEffect } from "react";
 import { ArticleContent } from "../components/article";
+import ServiceProcessCards from "../components/ServiceProcessCards";
 import { SplitLayout } from "../components/split-layout";
 import About from "./about";
+import ReactGA from "react-ga4";
+
 
 const serviceProcessCardsContent = [
   {
@@ -27,22 +29,12 @@ const serviceProcessCardsContent = [
   },
 ];
 
-const ServiceProcessCards = ({ data }) => {
-  return (
-    <div className="hover:bg-gray-100 dark:border-gray-700 dark:bg-gray-800 dark:hover:bg-gray-700 my-12 flex flex-col items-center rounded-lg border bg-white shadow-md md:max-w-xl md:flex-row">
-      <div className="h-96 w-full rounded-t-lg object-cover p-5 md:h-auto md:w-48 md:rounded-none md:rounded-l-lg">
-        <i className={data.iconClasses} />
-      </div>
-      <div className="flex flex-col justify-between p-4 leading-normal">
-        <h5 className="text-gray-900 mb-2 text-2xl font-bold tracking-tight dark:text-white">
-          {data.title}
-        </h5>
-      </div>
-    </div>
-  );
-};
+
 
 const HowWorks = () => {
+  useEffect(() => {
+    ReactGA.send({ hitType: "pageview", page: window.location.pathname });
+  }, []);
   return (
     <ArticleContent title="How does it work?">
       <h4>Easiest way to invest in cryptocurrencies.</h4>
